dict: fix list scrolling behavior
1) fix scrolling becoming stuck when rendering a new set of items
2) reset the scroll anchor when loading a new set of items

Root cause of (1) was the `scrollTop` and List `anchor` becoming out of sync when rendering a new set of items, causing the scroll to be applied entirely through the `translateY(offset)`. Since the `scrollTop` was not synced, in some situations the user would become unable to scroll without the virtual scrollbar.

Change (2) is not necessary for the bugfix but improves usability by reseting the scroll position when a new query is loaded.